rename library to vision3d
