+++ vision3d/detector/model.py
