+++ vision3d/detector/sparse_cnn.py
