+++ vision3d/ops/iou_nms.py
@@ -5,8 +5,8 @@ import torch
 from torchvision.ops import boxes as box_ops
 from torchvision.ops import nms  # BC-compat
 
-from pvrcnn import _C
-from pvrcnn._C import box_iou_rotated
+from vision3d import _C
+from vision3d._C import box_iou_rotated
 
 
 def box_iou_rotated_3d(boxes1, boxes2):
